requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dnspython==2.3.0
emoji==2.8.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
py-cord==2.4.1
pymongo==4.4.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
watchdog==3.0.0
yarl==1.9.2
